% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gangs.R
\docType{data}
\name{gangs}
\alias{gangs}
\title{Map of Non-Hispanic Gangs in South Los Angeles (2016)}
\format{Zipped KML file.}
\source{
Anonymous. Gangs of Los Angeles (2016). \url{https://goo.gl/7Ar1Aa} 
(Google My Maps, accessed 30 December 2016).
}
\description{
This map shows the non-Hispanic gangs present in South Los Angeles as of
December 2016.
}
\details{
The map comes from Google My Maps website and has been used to illustrate the
Wikipedia entry for the service.

The precise identity of the author of the map could not be determined.
}
\examples{
f <- system.file("extdata", "gangs.kml.zip", package = "tidykml")
kml_polygons(f)
}
\seealso{
LA Hood Maps. \emph{Instagram}.
\url{https://www.instagram.com/la_hood_maps/} (accessed 30 December 2016).
}

