% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kml_bounds.R
\name{kml_bounds}
\alias{kml_bounds}
\title{Find the boundaries of a KML file.}
\usage{
kml_bounds(x, ns = "d1", verbose = TRUE)
}
\arguments{
\item{x}{A KML source. See \link{kml_read}. \code{x} can also be a data frame
with two numeric variables named \code{longitude} and \code{latitude}.}

\item{ns}{The name of the namespace to extract from: defaults to \code{"d1"}.}

\item{verbose}{Whether to report invalid coordinates and/or altitudes below
sea level; defaults to \code{TRUE}. See \link{kml_coords}.}
}
\value{
A named numeric vector of four elements corresponding to the 
left, bottom, right and top values of the bounding box; 'left' and 'right' 
are the minimal and maximal longitudes; 'bottom' and 'top' are the minimal 
and maximal latitudes.
}
\description{
Find the boundaries of a KML file.
}
\examples{
# demo data: U.S. Civil War map
# see ?states for details
f <- system.file("extdata", "states.kml.zip", package = "tidykml")
kml_bounds(f)
}
\seealso{
\link{kml_coords}
}

