% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kml_coords.R
\name{kml_coords}
\alias{kml_coords}
\title{Extract KML coordinates (longitude, latitude or altitude).}
\usage{
kml_coords(x, coord, verbose = TRUE)
}
\arguments{
\item{x}{A character vector of KML coordinates, of the form
\code{"longitude,latitude[,altitude]"}.}

\item{coord}{Which coordinate to extract: either \code{1} (longitude), 
\code{2} (latitude) or \code{3} (altitude). The function also accepts
 \code{"lon"}, \code{"lat"} and \code{"alt"}.}

\item{verbose}{Whether to report invalid coordinates and/or negative
altitudes (below sea level); defaults to \code{TRUE}. See 'Note'.}
}
\value{
A numeric vector.

If \code{coord} has been set to return altitude and there are no altitude 
values in the KML coordinates, the function returns a numeric vector of
\code{NA} values of the same length as \code{x}.
}
\description{
Extract KML coordinates (longitude, latitude or altitude).
}
\note{
KML coordinates are \code{lon,lat[,alt]} tuples.

Invalid coordinates are longitudes outside of [-180, 180] (angular distance 
in degrees, relative to the Prime Meridian) and latitudes outside of 
[-90, 90] (angular distance in degrees, relative to the Equator).

Altitude is the distance from the earth's surface, in meters. Depending on
the altitude mode used in the KML source, altitude might be computed in 
absolute terms (relative to sea level), or from the actual ground elevation
of a particular location.
}
\examples{
# longitude
kml_coords("11.0,22.0,-99.0", 1)
# latitude
kml_coords("11.0,22.0,-99.0", 2)
# altitude
kml_coords("11.0,22.0,-99.0", "alt")
kml_coords("11.0,22.0,-99.0", "alt", verbose = FALSE)
}
\seealso{
Google Developers. KML Reference: <Camera> Element.
\url{https://developers.google.com/kml/documentation/kmlreference#camera}
}

