% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{kml_finalize}
\alias{kml_finalize}
\title{Finalize a KML tidy data frame}
\usage{
kml_finalize(x, folders, verbose = TRUE)
}
\arguments{
\item{x}{The KML data frame to tidy.}

\item{folders}{The number of folders in the data frame.}

\item{verbose}{Whether to report invalid coordinates and/or negative
altitudes (below sea level); defaults to \code{TRUE}.}
}
\description{
Reads the coordinates out of the \code{coordinates} variable, checks them,
and returns the data.
}
\keyword{internal}

