% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{kml_folders}
\alias{kml_folders}
\title{Extract KML Folders.}
\usage{
kml_folders(x, ns = "d1")
}
\arguments{
\item{x}{An XML document.}

\item{ns}{The name of the namespace to extract from; defaults to \code{"d1"}.}
}
\value{
A nodeset of Folders.
}
\description{
Extract KML Folders.
}
\seealso{
Google Developers. KML Reference: <Folder> Element.
\url{https://developers.google.com/kml/documentation/kmlreference#folder}
}
\keyword{internal}

