% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kml_info.R
\name{kml_info}
\alias{kml_info}
\title{Find the elements of a KML file.}
\usage{
kml_info(x, ns = "d1")
}
\arguments{
\item{x}{A KML source. See \link{kml_read}.}

\item{ns}{The name of the namespace to extract from: defaults to \code{"d1"}.}
}
\value{
A named numeric vector of five elements corresponding to the number
of Folders, Placemarks, LineStrings, Points and Polygons in the source.
}
\description{
Find the elements of a KML file.
}
\examples{
# demo data: U.S. Civil War map
# see ?states for details
f <- system.file("extdata", "states.kml.zip", package = "tidykml")
kml_info(f)
}

