% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{kml_placemarks}
\alias{kml_placemarks}
\title{Extract KML Placemarks containing a specific Geometry.}
\usage{
kml_placemarks(x, geometry, ns = "d1")
}
\arguments{
\item{x}{A nodeset of Folders.}

\item{geometry}{The name of the Geometry to subset on, e.g. \code{"Point"}.}

\item{ns}{The name of the namespace to extract from; defaults to \code{"d1"}.}
}
\value{
A nodeset of Placemarks.
}
\description{
Extract KML Placemarks containing a specific Geometry.
}
\seealso{
Google Developers. KML Reference: <Placemark> Element.
\url{https://developers.google.com/kml/documentation/kmlreference#placemark}
}
\keyword{internal}

