% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{kml_size}
\alias{kml_size}
\title{Find the number of coordinates in a KML file.}
\usage{
kml_size(x, ns = "d1")
}
\arguments{
\item{x}{A KML source. See \link{kml_read}.}

\item{ns}{The name of the namespace to extract from; defaults to \code{"d1"}.}
}
\value{
A named numeric vector of three elements containing the total number 
of coordinates, the total number of coordinates found in <outerBoundaryIs> 
elements (outer polygon boundaries), and the total number of coordinates 
found in <innerBoundaryIs> elements (inner polygon boundaries).
}
\description{
Find the number of coordinates in a KML file.
}
\keyword{internal}

