% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-states.R
\docType{data}
\name{states}
\alias{states}
\title{Map of U.S. Civil War}
\format{Zipped KML file.}
\source{
Anonymous. US Civil War. \url{https://goo.gl/rezvty}
(Google My Maps, accessed 30 December 2016).
}
\description{
This map shows the major battles of the U.S. Civil War, the states engaged in
the conflict and their status as of 1863, and the itinerary of Union General 
William T. Sherman's forces from Atlanta to Savannah, Georgia, in 1864.
}
\details{
The map comes from Google My Maps website and has been used to illustrate the
Wikipedia entry for the service.

The precise identity of the author of the map could not be determined.
}
\examples{
f <- system.file("extdata", "states.kml.zip", package = "tidykml")
kml_points(f)
}
\seealso{
Wikipedia. Google My Maps. \emph{Wikipedia, The Free Encyclopedia}.
\url{https://en.wikipedia.org/wiki/Google_My_Maps} (revision 732328417,
accessed 30 December 2016).
}

