% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidykml.R
\docType{package}
\name{tidykml}
\alias{tidykml}
\alias{tidykml-package}
\title{Functions to turn KML (Keyhole Markup Language) files into tidy data frames.}
\description{
The \code{tidykml} package reads selected elements and values from KML files,
such as those produced by Google My Maps, and turns them into tidy data 
frames, intended for use with packages like \link[dplyr:dplyr]{dplyr} and 
\link[ggplot2:ggplot2]{ggplot2}.
}
\details{
See the README file of the package for further details and a few examples:
\url{https://github.com/briatte/tidykml}
}
\seealso{
Google Developers. KML Reference.
\url{https://developers.google.com/kml/documentation/kmlreference}
}

