% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kml_read.R
\name{kml_read}
\alias{kml_read}
\title{Read a KML file.}
\usage{
kml_read(x, ...)
}
\arguments{
\item{x}{A string, a connection, or a raw vector.
All inputs accepted by \link[xml2:read_xml]{read_xml} are also accepted, as 
long as they are valid KML. This means that URLs and local compressed files,
such as \code{.zip} files, are also supported. 
See \link[xml2:read_xml]{read_xml} for details.}

\item{...}{Arguments passed to \link[xml2:read_xml]{read_xml}, such as 
\code{encoding} or \code{base_url}.
See \link[xml2:read_xml]{read_xml} for details.}
}
\value{
A nodeset.
}
\description{
Read a KML file.
}
\examples{
# demo data: U.S. Civil War map
# see ?states for details
f <- system.file("extdata", "states.kml.zip", package = "tidykml")
kml_read(f)
}
\seealso{
Google Developers. KML Reference.
\url{https://developers.google.com/kml/documentation/kmlreference}
}

